refactor : complete move of commands to simpleaf_commands module
